pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wrds-mcp"
version = "0.1.0"
description = "MCP server for WRDS financial data access — credit analysis tools"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=3.0",
    "wrds>=3.5",
    "pandas>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/wrds_mcp --cov-report=term-missing"
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/wrds_mcp"]

[project.scripts]
wrds-mcp = "wrds_mcp.server:main"
